@@ -1,6 +1,6 @@
 package com.kentvu.quanlyquy
 
-import android.support.test.rule.ActivityTestRule
+import androidx.test.rule.ActivityTestRule
 import org.junit.Rule
 
 class AppRunner {
